feat: Dealer와 Player의  공통 행위 추상화를 위한 Participant 추가
