+++ src/main/java/blackjack/domain/participant/Dealer.java
@@ -2,52 +2,32 @@ package blackjack.domain.participant;
 
 import blackjack.domain.GameResult;
 import blackjack.domain.card.Card;
-import blackjack.domain.card.Cards;
 import java.util.List;
 
-public class Dealer {
+public class Dealer extends Participant {
 
     private static final String DEALER_NAME = "딜러";
     private static final int DRAWABLE_LIMIT_VALUE = 16;
-    private static final int PLAYING_STANDARD = 21;
-
-    private final Name name;
-    private final Cards cards;
 
     public Dealer(List<Card> cards) {
-        this.name = new Name(DEALER_NAME);
-        this.cards = new Cards(cards);
+        super(DEALER_NAME, cards);
     }
 
-    public int getTotalScore() {
-        return cards.calculateTotalScore();
-    }
-
+    @Override
     public boolean isDrawable() {
         return cards.calculateTotalScore() <= DRAWABLE_LIMIT_VALUE;
     }
 
-    public void combine(Card card) {
-        cards.combine(card);
-    }
-
+    @Override
     public GameResult decideResult(int playerScore) {
-        if (playerScore > PLAYING_STANDARD) {
+        if (playerScore > BLACKJACK_SCORE) {
             return GameResult.WIN;
         }
 
-        if (getTotalScore() > PLAYING_STANDARD) {
+        if (getTotalScore() > BLACKJACK_SCORE) {
             return GameResult.LOSE;
         }
 
         return GameResult.of(getTotalScore() - playerScore);
-    }
-
-    public String getName() {
-        return name.getValue();
-    }
-
-    public List<Card> getCards() {
-        return cards.getValue();
     }
 }
